CB-1261 replace systemctl for ycloud sssd

+++ orchestrator-salt/src/main/resources/salt/salt/sssd/ipa.sls
@@ -52,6 +52,38 @@ restart-sssd-if-reconfigured:
     - user: root
     - group: root
 
+{% if salt['file.directory_exists']('/yarn-private') %}
+
+backup-systemctl:
+  file.copy:
+    - name: /usr/bin/systemctl.bak
+    - source: /usr/bin/systemctl.orig
+
+replace_systemctl:
+  file.managed:
+    - source: salt://sssd/ycloud/systemctl
+    - name: /usr/bin/systemctl.orig
+    - user: root
+    - group: root
+    - mode: 755
+
+force_restart_ssd:
+  cmd.run:
+    - name: runuser -l root -c 'systemctl restart sssd'
+    - runas: root
+    - require:
+        - file: replace_systemctl
+
+restore_systemctl:
+  file.copy:
+    - name: /usr/bin/systemctl.orig
+    - source: /usr/bin/systemctl.bak
+    - user: root
+    - group: root
+    - mode: 755
+    - force: True
+{% endif %}
+
 include:
     - sssd.ssh
 
@@ -69,7 +101,12 @@ create_cm_keytab_generation_script:
 
 generate_cm_freeipa_keytab:
   cmd.run:
+{% if not salt['file.directory_exists']('/yarn-private') %}
     - name: sh /opt/salt/scripts/generate_cm_keytab.sh 2>&1 | tee -a /var/log/generate_cm_keytab.log && exit ${PIPESTATUS[0]}
+{% else %}
+    - name: newgrp root && sh /opt/salt/scripts/generate_cm_keytab.sh 2>&1 | tee -a /var/log/generate_cm_keytab.log && exit ${PIPESTATUS[0]}
+{% endif %}
+    - runas: root
     - env:
         - password: {{salt['pillar.get']('sssd-ipa:password')}}
     - unless: ls /etc/cloudera-scm-server/cmf.keytab
